feat: correctly order the games
This required a larger refactoring to use arrays in firebase which now
support querying by element. This allows us to change the 'playerIds'
from the silly map of elements (like `{ playerOne: true, playerTwo:
true`) to just a simple array of ids.

Also updated the tests and (as it was necessary) updated some
dependencies and added code to convert the playerIds.